Add INT4 fake-quant, BF16 autocast, full PTQ results, attn-quant handoff
Quantization infrastructure:
- quantize_torchao: INT4 symmetric per-group fake-quant (_apply_int4_weight_fake_quant),
  fix model_size_mb for LinearActivationQuantizedTensor (int8dq), handle all 3 param types
- eval_ptq.py: --use_autocast flag (BF16 stacked on any scheme), build_run_name suffix
- models/retr.py: fix BF16 dtype cast in mask logit assignment (line 245)

Experiment results (P2S1 CPU, 9 runs):
- int8wo decoder/transformer/all: AP=42.73 (−0.05), IoU=74.41
- int8dq all: AP=42.66 (−0.12), IoU=74.40
- int8dq+bf16 all: AP=41.07 (−1.71) — not viable
- int4fq_g128/g64/g32 all: AP=42.93/42.24/41.72

Documentation:
- README: full PTQ results table, schemes, literature threshold table, attention matrix
  sizing analysis (encoder 512×512 → online softmax required), hardware sizing note
- docs/experiments.md: scheme name updates (int4fq_*), mslk note
- HANDOFF_ATTN_QUANT.md: implementation guide for attention bmm fake-quant (next step)
- README: placeholder rows + citations for pending attn-bmm INT8 experiments

+++ src/models/retr.py
@@ -242,7 +242,7 @@ class RETR(nn.Module):
                     final_masks[q][mask_thresh] = 1
                     final_mask[mask_thresh] = 1
                     positive = current_mask[0] > 0
-                    final_logits[positive] = current_mask[0][positive]
+                    final_logits[positive] = current_mask[0][positive].to(final_logits.dtype)
 
                 predictions.append(final_mask)
                 final_mask_logits.append(final_logits)
